Fix height scaling when using software::scaling
The output height of an image was incorrectly passed to srcSliceH, which
expects the _source_ image height rather than the destination image
height.

The current logic results in an error or otherwise incorrect behavior when the
source image height was different from the destination image height,
e.g. "Slice parameters 0, 1072 are invalid"